Optimized data loading and storing (changed vectors of floats to vectors of int16_t, now position of vertex is calculated in vertex shader).

+++ src/vbo.cpp
@@ -17,7 +17,7 @@ VBO::VBO(std::vector<float>* vertices, GLsizeiptr size)
     glBufferData(GL_ARRAY_BUFFER, size, &vertices->front(), GL_STATIC_DRAW);
 }
 
-VBO::VBO(std::vector<double>* vertices, GLsizeiptr size)
+VBO::VBO(std::vector<int16_t>* vertices, GLsizeiptr size)
 {
     glGenBuffers(1, &id);
     glBindBuffer(GL_ARRAY_BUFFER, id);
